Factor out invalid source request testing
